Update import transaction to include signatures under Sign, add a box shadow
