+++ src/app/(sidebar)/transaction/import/components/SignStepSignatureContext.tsx
@@ -3,13 +3,8 @@ import {
   Transaction,
   TransactionBuilder,
 } from "@stellar/stellar-sdk";
-import { Notification } from "@stellar/design-system";
 
 import { useStore } from "@/store/useStore";
-import {
-  getTxSignatureCompleteness,
-  TxSignatureCompleteness,
-} from "@/helpers/checkRequiredSignatures";
 
 import { Box } from "@/components/layout/Box";
 import { Signatures } from "@/app/(sidebar)/transaction/components/Signatures";
@@ -20,41 +15,10 @@ type Props = {
   parsedTxType?: "classic" | "soroban" | null;
 };
 
-type AlertVariant = "primary" | "success" | "warning" | "error";
-
 const hasAnySignature = (tx: Transaction | FeeBumpTransaction): boolean =>
   tx instanceof FeeBumpTransaction
     ? tx.signatures.length > 0 || tx.innerTransaction.signatures.length > 0
     : tx.signatures.length > 0;
-
-const getContextMessage = (
-  completeness: TxSignatureCompleteness,
-): { variant: AlertVariant; message: string } => {
-  if (completeness.hasInvalid) {
-    return {
-      variant: "error",
-      message:
-        "This transaction carries invalid signature(s) that won’t be accepted at submission. Review the signatures below before signing.",
-    };
-  }
-
-  if (
-    completeness.missingSigners.length > 0 ||
-    completeness.hasUnrecognizedSigners
-  ) {
-    return {
-      variant: "warning",
-      message:
-        "This transaction already carries signature(s), but offline checks can’t confirm every required signer is covered — a multisig account may be signed by cosigners. If you’re a required signer, add yours below; otherwise you can continue to submit and let the network verify.",
-    };
-  }
-
-  return {
-    variant: "success",
-    message:
-      "This transaction already has every signature that can be verified offline.",
-  };
-};
 
 export const SignStepSignatureContext = ({ xdr, parsedTxType }: Props) => {
   const { network } = useStore();
@@ -74,14 +38,8 @@ export const SignStepSignatureContext = ({ xdr, parsedTxType }: Props) => {
     return null;
   }
 
-  const { variant, message } = getContextMessage(
-    getTxSignatureCompleteness(tx),
-  );
-
   return (
     <Box gap="md">
-      <Notification variant={variant} title={message} />
-
       <Signatures tx={tx} parsedTxType={parsedTxType} />
     </Box>
   );
